fix modal props in header +  component logic in app.js
